refactor: ensure single-responsibility principle for handler
