+++ src/handlers/users/create.js
@@ -1,27 +1,27 @@
 import ValidationError from '../../validators/errors/validation-error'
-import validate from '../../validators/users/create'
+import create from '../../engines/users/create'
 
 function createUser(req, res, db) {
-  const validationResults = validate(req)
-  if (validationResults instanceof ValidationError) {
-    res.status(400)
-    res.set('Content-Type', 'application/json')
-    return res.json({ message: validationResults.message })
-  }
-  db.index({
-    index: process.env.ELASTICSEARCH_INDEX,
-    type: 'user',
-    body: req.body
-  })
-    .then(result => {
-      res.status(201)
-      res.set('Content-Type', 'text/plain')
-      res.send(result._id)
-    })
+  create(req, db)
+    .then(
+      result => {
+        res.status(201)
+        res.set('Content-Type', 'text/plain')
+        return res.send(result._id)
+      },
+      err => {
+        if (err instanceof ValidationError) {
+          res.status(400)
+          res.set('Content-Type', 'application/json')
+          return res.json({ message: err.message })
+        }
+        return undefined
+      }
+    )
     .catch(() => {
       res.status(500)
       res.set('Content-Type', 'application/json')
-      res.json({ message: 'Internal Server Error' })
+      return res.json({ message: 'Internal Server Error' })
     })
 }
